+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Give `<Button>` same size/padding/border/margin as inputs so they line up nicely side by side
+- change how `complimentaryColor` works
 
 ## [0.17.1]
 ### Changed