+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+*N/A*
+
+## [0.18.0]
 ### Added
 - `color?: Colors` prop added to `<Notification>`
 - React hook `useClickAway`
@@ -253,7 +256,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First version published
 
-[unreleased]: https://github.com/chadlavi/clear/compare/v0.17.1...HEAD
+[unreleased]: https://github.com/chadlavi/clear/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/chadlavi/clear/compare/v0.17.1...v0.18.0
 [0.17.1]: https://github.com/chadlavi/clear/compare/v0.16.1...v0.17.1
 [0.16.1]: https://github.com/chadlavi/clear/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/chadlavi/clear/compare/v0.15.0...v0.16.0